# docs/week1/backend/openapi-draft.yaml
openapi: 3.1.0
info:
  title: Agent Audit Receipt API
  version: 0.1.0
  summary: Draft API for verification and receipt retrieval.
servers:
  - url: https://api.agentbanking.dev
paths:
  /verify:
    post:
      operationId: createVerification
      summary: Verify an agent run and issue a receipt.
      tags: [Verification]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyRequest'
      responses:
        '201':
          description: Verification accepted and receipt issued.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyResponse'
        '400':
          description: Invalid request payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid authentication.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Operator not allowed to verify this agent/request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Duplicate verification request detected by idempotency key/request_id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Verification processed but integrity checks failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntegrityFailureResponse'
        '429':
          description: Too many requests.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal verification pipeline error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Dependent proof service unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /verify/offline:
    post:
      operationId: verifyReceiptOffline
      summary: Verify issued receipt/report pair without re-running online pipeline.
      tags: [Verification]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OfflineVerifyRequest'
      responses:
        '200':
          description: Offline verification passed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OfflineVerifyResult'
        '400':
          description: Invalid offline verification payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid authentication.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Offline verification failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntegrityFailureResponse'
        '500':
          description: Internal verification error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /receipts/{receiptId}:
    get:
      operationId: getReceiptById
      summary: Retrieve an issued receipt.
      tags: [Receipts]
      parameters:
        - in: path
          name: receiptId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Receipt found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Receipt'
        '401':
          description: Missing or invalid authentication.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Operator not allowed to access this receipt.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Receipt not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too many requests.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal retrieval error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    VerifyRequest:
      type: object
      required:
        - request_id
        - agent_id
        - operator_id
        - policy_version
        - strategy_hash
        - model_hash
        - evidence_refs
      properties:
        request_id:
          type: string
        agent_id:
          type: string
        operator_id:
          type: string
        policy_version:
          type: string
        strategy_hash:
          type: string
        model_hash:
          type: string
        evidence_refs:
          type: array
          items:
            $ref: '#/components/schemas/EvidenceRef'
    VerifyResponse:
      type: object
      required: [receipt, audit_report]
      properties:
        receipt:
          $ref: '#/components/schemas/Receipt'
        audit_report:
          $ref: '#/components/schemas/AuditReport'
    OfflineVerifyRequest:
      type: object
      required: [receipt, audit_report]
      properties:
        receipt:
          $ref: '#/components/schemas/Receipt'
        audit_report:
          $ref: '#/components/schemas/AuditReport'
        public_key_pem:
          type: string
          description: Optional Ed25519 public key PEM for strict signature verification.
        strict_signature:
          type: boolean
          default: false
    OfflineVerifyCheck:
      type: object
      required: [name, ok, code, severity, message]
      properties:
        name:
          type: string
        ok:
          type: boolean
        code:
          type: string
        severity:
          type: string
          enum: [low, medium, high, critical]
        message:
          type: string
    OfflineVerifyResult:
      type: object
      required:
        - correlation_id
        - verification_result
        - failed_codes
        - warnings
        - checks
      properties:
        correlation_id:
          type: string
          format: uuid
        verification_result:
          type: string
          enum: [PASS, FAIL]
        failed_codes:
          type: array
          items:
            type: string
        warnings:
          type: array
          items:
            type: string
        checks:
          type: array
          items:
            $ref: '#/components/schemas/OfflineVerifyCheck'
    Receipt:
      type: object
      required:
        - receipt_id
        - report_id
        - request_id
        - agent_id
        - operator_id
        - schema_version
        - issued_at
        - verification_result
        - retention_until
        - verification_endpoint
      properties:
        receipt_id:
          type: string
          format: uuid
        report_id:
          type: string
        request_id:
          type: string
        agent_id:
          type: string
        operator_id:
          type: string
        schema_version:
          type: string
          example: 1.0.0
        issued_at:
          type: string
          format: date-time
        verification_result:
          type: string
          enum: [pass, warning, fail]
        policy_version:
          type: string
        report_digest:
          $ref: '#/components/schemas/Digest'
        signature:
          type: object
          properties:
            format:
              type: string
              enum: [jws-compact, jws-json]
            alg:
              type: string
              enum: [EdDSA, ES256]
            kid:
              type: string
            value:
              type: string
        timestamp_proof:
          type: object
          properties:
            tsa_name:
              type: string
            gen_time:
              type: string
              format: date-time
            token_b64:
              type: string
        transparency_proof:
          type: object
          properties:
            log_id:
              type: string
            entry_id:
              type: string
            leaf_hash:
              type: string
            root_hash:
              type: string
            tree_size:
              type: integer
            inclusion_proof:
              type: array
              items:
                type: string
        evidence_refs:
          type: array
          items:
            $ref: '#/components/schemas/EvidenceRef'
        retention_until:
          type: string
          format: date-time
        verification_endpoint:
          type: string
          format: uri
        links:
          type: object
          properties:
            verify_url:
              type: string
              format: uri
            report_url:
              type: string
              format: uri
    AuditReport:
      type: object
      required: [report_id, summary, findings]
      properties:
        report_id:
          type: string
        summary:
          type: string
        findings:
          type: array
          items:
            type: object
            properties:
              code:
                type: string
              severity:
                type: string
                enum: [info, low, medium, high, critical]
              message:
                type: string
    EvidenceRef:
      type: object
      required: [kind, uri, digest]
      properties:
        kind:
          type: string
          enum: [input, execution_log, output, policy_snapshot]
        uri:
          type: string
          format: uri
        digest:
          $ref: '#/components/schemas/Digest'
    Digest:
      type: object
      required: [alg, value]
      properties:
        alg:
          type: string
          enum: [sha-256, sha-512]
        value:
          type: string
    ErrorResponse:
      type: object
      required:
        - code
        - message
        - severity
        - category
        - retryable
        - correlation_id
      properties:
        code:
          type: string
          enum:
            - INVALID_REQUEST
            - UNAUTHORIZED
            - FORBIDDEN
            - DUPLICATE_REQUEST
            - INTEGRITY_CHECK_FAILED
            - SIGNATURE_VERIFICATION_FAILED
            - DIGEST_MISMATCH
            - TIMESTAMP_PROOF_INVALID
            - TRANSPARENCY_PROOF_INVALID
            - REQUEST_REPLAY_DETECTED
            - RATE_LIMITED
            - INTERNAL_ERROR
            - PROOF_SERVICE_UNAVAILABLE
            - RECEIPT_NOT_FOUND
        message:
          type: string
        severity:
          type: string
          enum: [info, low, medium, high, critical]
        category:
          type: string
          enum: [validation, authz, conflict, integrity, rate_limit, system, not_found]
        retryable:
          type: boolean
        correlation_id:
          type: string
          format: uuid
        details:
          type: object
          additionalProperties: true
        remediation:
          type: string
    IntegrityFailureItem:
      type: object
      required:
        - check
        - reason_code
        - message
        - severity
        - retryable
      properties:
        check:
          type: string
          enum:
            - signature
            - digest
            - timestamp_proof
            - transparency_proof
            - request_replay
        reason_code:
          type: string
          enum:
            - SIGNATURE_INVALID
            - DIGEST_MISMATCH
            - TIMESTAMP_TOKEN_INVALID
            - TIMESTAMP_TIME_SKEW
            - TRANSPARENCY_INCLUSION_MISSING
            - TRANSPARENCY_ROOT_MISMATCH
            - REQUEST_REPLAY_DETECTED
        message:
          type: string
        severity:
          type: string
          enum: [low, medium, high, critical]
        retryable:
          type: boolean
        evidence_uri:
          type: string
          format: uri
        observed_at:
          type: string
          format: date-time
    IntegrityFailureResult:
      type: object
      required:
        - verification_result
        - failed_checks
      properties:
        verification_result:
          type: string
          enum: [fail]
        failed_checks:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/IntegrityFailureItem'
    IntegrityFailureResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          required: [integrity_result]
          properties:
            integrity_result:
              $ref: '#/components/schemas/IntegrityFailureResult'